Event based Asynchronous Processor
Power is one of the major concern of design engineers, using event-based sampling technique the power consumption can be reduced. In the project, analog to digital conversion is done using level crossing techniques and then a Low Pass Filter is implemented asynchronously on an FPGA with analog delay blocks(made on hardware). The overall power consumption of the project was 0.5W which can be further reduced.